10064: WIP removing feature flags
